// types.ts
export interface PubMedArticle {
    uid: string;
    title: string;
    source: string; // Journal
    pubdate: string;
    authors: Array<{ name: string }>;
    volume?: string;
    url: string;
}